[Proposal] GLUE processors included in library
